tests/wpt/web-platform-tests/workers/baseurl/alpha/importScripts.html

<!doctype html>
<meta charset=utf-8>
<title>Base URL in workers: importScripts</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(function() {
  var worker = new Worker("../beta/importScripts.py");
  worker.onmessage = this.step_func_done(function(e) {
    assert_equals(e.data, "gamma");
  });
  worker.onerror = this.unreached_func("Got error event");
});
</script>
