third_party/WebKit/LayoutTests/platform/win/external/wpt/workers/baseurl/alpha/importScripts-expected.txt

This is a testharness.js-based test.
FAIL Base URL in workers: importScripts assert_equals: expected "gamma" but got "beta"
Harness: the test ran to completion.
